+++ backend/migrations/20240108130114_add_oidc.up.sql
@@ -1,0 +1,2 @@
+-- Add up migration script here
+ALTER TABLE token ADD COLUMN IF NOT EXISTS job UUID;